+++ src/torchmetrics/classification/roc.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import Any, List, Optional, Tuple, Union
+from typing import Any, List, Optional, Sequence, Tuple, Union
 
 from torch import Tensor
 from typing_extensions import Literal
@@ -21,6 +21,14 @@ from torchmetrics.classification.precision_recall_curve import (
     MulticlassPrecisionRecallCurve,
     MultilabelPrecisionRecallCurve,
 )
+from torchmetrics.functional.classification.auroc import (
+    _binary_auroc_arg_validation,
+    _binary_auroc_compute,
+    _multiclass_auroc_arg_validation,
+    _multiclass_auroc_compute,
+    _multilabel_auroc_arg_validation,
+    _multilabel_auroc_compute,
+)
 from torchmetrics.functional.classification.roc import (
     _binary_roc_compute,
     _multiclass_roc_compute,
@@ -29,12 +37,18 @@ from torchmetrics.functional.classification.roc import (
 from torchmetrics.metric import Metric
 from torchmetrics.utilities.data import dim_zero_cat
 from torchmetrics.utilities.enums import ClassificationTask
+from torchmetrics.utilities.imports import _MATPLOTLIB_AVAILABLE
+from torchmetrics.utilities.plot import _AX_TYPE, _PLOT_OUT_TYPE, plot_binary_roc_curve
+
+if not _MATPLOTLIB_AVAILABLE:
+    __doctest_skip__ = ["BinaryROC.plot"]
 
 
 class BinaryROC(BinaryPrecisionRecallCurve):
-    r"""Compute the Receiver Operating Characteristic (ROC) for binary tasks. The curve consist of multiple pairs of
-    true positive rate (TPR) and false positive rate (FPR) values evaluated at different thresholds, such that the
-    tradeoff between the two values can be seen.
+    r"""Compute the Receiver Operating Characteristic (ROC) for binary tasks.
+
+    The curve consist of multiple pairs of true positive rate (TPR) and false positive rate (FPR) values evaluated at
+    different thresholds, such that the tradeoff between the two values can be seen.
 
     As input to ``forward`` and ``update`` the metric accepts the following input:
 
@@ -107,11 +121,59 @@ class BinaryROC(BinaryPrecisionRecallCurve):
         state = [dim_zero_cat(self.preds), dim_zero_cat(self.target)] if self.thresholds is None else self.confmat
         return _binary_roc_compute(state, self.thresholds)
 
+    def __compute_auroc(self) -> Tensor:
+        """Computes Area under ROC curve from BinaryAUROC metric to show AUROC value together with ROC plot."""
+        state = [dim_zero_cat(self.preds), dim_zero_cat(self.target)] if self.thresholds is None else self.confmat
+        return _binary_auroc_compute(state, self.thresholds, max_fpr=None)
+
+    def plot(
+        self,
+        fpr: Optional[Union[Tensor, Sequence[Tensor]]] = None,
+        tpr: Optional[Union[Tensor, Sequence[Tensor]]] = None,
+        ax: Optional[_AX_TYPE] = None,
+        name: Optional[str] = None,
+    ) -> _PLOT_OUT_TYPE:
+        """Plot a single or multiple values from the metric.
+
+        Args:
+            fpr: False Postive Rate provided by calling `metric.forward` or `metric.compute`
+            tpr: True Postive Rate provided by calling `metric.forward` or `metric.compute`
+                If no value is provided, will automatically call `metric.compute` and plot that result.
+            ax: An matplotlib axis object. If provided will add plot to that axis
+            name: Custom name to describe the classifier
+
+        Returns:
+            Figure and Axes object
+
+        Raises:
+            ModuleNotFoundError:
+                If `matplotlib` is not installed
+
+        .. plot::
+            :scale: 75
+
+            >>> from torch import randn, randint
+            >>> import torch.nn.functional as F
+            >>> from torchmetrics.classification import BinaryROC
+            >>> preds = F.softmax(randn(20, 2), dim=1)
+            >>> target = randint(2, (20,))
+            >>> metric = BinaryROC()
+            >>> metric.update(preds[:, 1], target)
+            >>> fig_, ax_ = metric.plot()
+        """
+        if fpr is None or tpr is None:
+            fpr, tpr, _ = self.compute()
+        roc_auc = self.__compute_auroc()
+        name = self.__class__.__name__ if name is None else name
+        fig, ax = plot_binary_roc_curve(tpr, fpr, ax=ax, roc_auc=roc_auc, name=name)
+        return fig, ax
+
 
 class MulticlassROC(MulticlassPrecisionRecallCurve):
-    r"""Compute the Receiver Operating Characteristic (ROC) for binary tasks. The curve consist of multiple pairs of
-    true positive rate (TPR) and false positive rate (FPR) values evaluated at different thresholds, such that the
-    tradeoff between the two values can be seen.
+    r"""Compute the Receiver Operating Characteristic (ROC) for binary tasks.
+
+    The curve consist of multiple pairs of true positive rate (TPR) and false positive rate (FPR) values evaluated at
+    different thresholds, such that the tradeoff between the two values can be seen.
 
     As input to ``forward`` and ``update`` the metric accepts the following input:
 
@@ -211,9 +273,10 @@ class MulticlassROC(MulticlassPrecisionRecallCurve):
 
 
 class MultilabelROC(MultilabelPrecisionRecallCurve):
-    r"""Compute the Receiver Operating Characteristic (ROC) for binary tasks. The curve consist of multiple pairs of
-    true positive rate (TPR) and false positive rate (FPR) values evaluated at different thresholds, such that the
-    tradeoff between the two values can be seen.
+    r"""Compute the Receiver Operating Characteristic (ROC) for binary tasks.
+
+    The curve consist of multiple pairs of true positive rate (TPR) and false positive rate (FPR) values evaluated at
+    different thresholds, such that the tradeoff between the two values can be seen.
 
     As input to ``forward`` and ``update`` the metric accepts the following input:
 
@@ -315,9 +378,10 @@ class MultilabelROC(MultilabelPrecisionRecallCurve):
 
 
 class ROC:
-    r"""Compute the Receiver Operating Characteristic (ROC). The curve consist of multiple pairs of true positive
-    rate (TPR) and false positive rate (FPR) values evaluated at different thresholds, such that the tradeoff
-    between the two values can be seen.
+    r"""Compute the Receiver Operating Characteristic (ROC).
+
+    The curve consist of multiple pairs of true positive rate (TPR) and false positive rate (FPR) values evaluated at
+    different thresholds, such that the tradeoff between the two values can be seen.
 
     This function is a simple wrapper to get the task specific versions of this metric, which is done by setting the
     ``task`` argument to either ``'binary'``, ``'multiclass'`` or ``multilabel``. See the documentation of