+++ src/torchmetrics/functional/detection/panoptic_quality.py
@@ -42,7 +42,7 @@ def _to_tuple(t: Tensor) -> Tuple:
 
 
 def _get_color_areas(img: Tensor) -> Dict[Tuple, Tensor]:
-    """Counts all color occurrences.
+    """Count all color occurrences.
 
     Args:
         img: the image tensor containing the colored pixels.
@@ -85,7 +85,7 @@ def _validate_categories(things: Set[int], stuffs: Set[int]) -> None:
 
 
 def _validate_inputs(preds: Tensor, target: torch.Tensor) -> None:
-    """Validates the shapes of prediction and target tensors.
+    """Validate the shapes of prediction and target tensors.
 
     Args:
         preds: the prediction tensor
@@ -118,7 +118,7 @@ def _get_void_color(things: Set[int], stuffs: Set[int]) -> Tuple[int, int]:
 
 
 def _get_category_id_to_continuous_id(things: Set[int], stuffs: Set[int]) -> Dict[int, int]:
-    """Converts original IDs to continuous IDs.
+    """Convert original IDs to continuous IDs.
 
     Args:
         things: all unique ids for things classes
@@ -138,7 +138,7 @@ def _get_category_id_to_continuous_id(things: Set[int], stuffs: Set[int]) -> Dic
 
 
 def _isin(arr: Tensor, values: List) -> Tensor:
-    """Basic implementation of torch.isin to support pre 0.10 version.
+    """Check if all values of an arr are in another array. Implementation of torch.isin to support pre 0.10 version.
 
     Args:
         arr: the torch tensor to check for availabilities
@@ -300,7 +300,7 @@ def panoptic_quality(
     stuffs: Set[int],
     allow_unknown_preds_category: bool = False,
 ) -> Tensor:
-    r"""Computes the `Panoptic Quality`_ for panoptic segmentations.
+    r"""Compute `Panoptic Quality`_ for panoptic segmentations.
 
     .. math::
         PQ = \frac{IOU}{TP + 0.5 FP + 0.5 FN}