+++ src/torchmetrics/functional/multimodal/clip_score.py
@@ -105,7 +105,7 @@ def clip_score(
         "openai/clip-vit-large-patch14",
     ] = "openai/clip-vit-large-patch14",
 ) -> Tensor:
-    r"""Calculates `CLIP Score`_ which is a text-to-image similarity metric.
+    r"""Calculate `CLIP Score`_ which is a text-to-image similarity metric.
 
     CLIP is a reference free metric that can be used to evaluate the correlation between a generated caption for an
     image and the actual content of the image. It has been found to be highly correlated with human judgement. The