+++ src/torchmetrics/functional/text/bleu.py
@@ -24,7 +24,7 @@ from torch import Tensor, tensor
 
 
 def _count_ngram(ngram_input_list: Sequence[str], n_gram: int) -> Counter:
-    """Counting how many times each word appears in a given text with ngram.
+    """Count how many times each word appears in a given text with ngram.
 
     Args:
         ngram_input_list: A list of translated text or reference texts