+++ src/torchmetrics/image/ergas.py
@@ -29,9 +29,10 @@ if not _MATPLOTLIB_AVAILABLE:
 
 
 class ErrorRelativeGlobalDimensionlessSynthesis(Metric):
-    """Calculate `Relative dimensionless global error synthesis`_ (ERGAS) is used to calculate the accuracy of Pan
-    sharpened image considering normalized average error of each band of the result image
-    (ErrorRelativeGlobalDimensionlessSynthesis).
+    """Calculate `Relative dimensionless global error synthesis`_ (ERGAS).
+
+    This metric is used to calculate the accuracy of Pan sharpened image considering normalized average error of each
+    band of the result image.
 
     As input to ``forward`` and ``update`` the metric accepts the following input