+++ src/torchmetrics/utilities/exceptions.py
@@ -17,5 +17,5 @@ class TorchMetricsUserError(Exception):
     """Error used to inform users of a wrong combination of Metric API calls."""
 
 
-class MisConfigurationError(Exception):
-    """Exception used to inform users of misuse of torchmetrics."""
+class TorchMetricsUserWarning(Warning):
+    """Error used to inform users of specific warnings due to the torchmetrics API."""