+++ tests/unittests/bases/test_aggregation.py
@@ -3,7 +3,8 @@ import pytest
 import torch
 
 from torchmetrics.aggregation import CatMetric, MaxMetric, MeanMetric, MinMetric, SumMetric
-from unittests.helpers.testers import BATCH_SIZE, NUM_BATCHES, MetricTester
+from unittests import BATCH_SIZE, NUM_BATCHES
+from unittests.helpers.testers import MetricTester
 
 
 def compare_mean(values, weights):